refactor : move database-to-excel in a separate module
