+++ app/src/main/AndroidManifest.xml
@@ -30,16 +30,6 @@
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
-
-        <provider
-            android:name=".ui.exportToExcel.TrainInventoryFileProvider"
-            android:authorities="com.canli.oya.traininventoryroom.fileprovider"
-            android:exported="false"
-            android:grantUriPermissions="true">
-            <meta-data
-                android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/file_paths"/>
-        </provider>
     </application>
 
 </manifest>
