+++ databasetoexcel/src/main/AndroidManifest.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.canlioya.databasetoexcel">
+
+    <application>
+        <provider
+            android:name="com.canlioya.databasetoexcel.TrainInventoryFileProvider"
+            android:authorities="com.canli.oya.traininventoryroom.fileprovider"
+            android:exported="false"
+            android:grantUriPermissions="true">
+            <meta-data
+                android:name="android.support.FILE_PROVIDER_PATHS"
+                android:resource="@xml/file_paths"/>
+        </provider>
+    </application>
+
+</manifest>
