+++ databasetoexcel/src/main/java/com/canlioya/databasetoexcel/TrainInventoryFileProvider.kt
@@ -1,4 +1,4 @@
-package com.canli.oya.traininventoryroom.ui.exportToExcel
+package com.canlioya.databasetoexcel
 
 import androidx.core.content.FileProvider
