+++ settings.gradle
@@ -2,3 +2,4 @@ include ':app'
 include ':core'
 include ':local'
 include ':testresources'
+include ':databasetoexcel'
